requirements.txt
Flask==3.0.3

Flask-Cors==4.0.1             
Flask-SQLAlchemy==3.1.1       
Flask-Migrate==4.0.7         
Flask-RESTful==0.3.10        
Flask-JWT-Extended==4.6.0    
Flask-Swagger-UI==4.11.1  
Flask-Marshmallow==1.3.0


bcrypt==4.2.0                 
itsdangerous==2.2.0     
marshmallow==3.21.1          
marshmallow-sqlalchemy==0.29.0
PyMySQL==1.1.1
cryptography>=41.0

python-dotenv==1.0.1        
requests==2.32.3          

Flasgger==0.9.7.1

PyMuPDF==1.24.10       
ebooklib==0.18
Pillow==10.4.0
